requirements.txt
# More flexible versioning
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
firebase-admin>=6.0.0

# LangChain ecosystem - using compatible ranges
langchain>=0.1.0,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.0.10,<1.0.0
langchain-google-genai>=1.0.0,<3.0.0
google-generativeai
langchain-chroma>=0.1.0,<1.0.0

# ML/AI packages
faiss-cpu>=1.7.0
numpy>=1.24.0
sentence-transformers>=2.0.0
pillow>=10.0.0

# Utilities
requests>=2.30.0
websockets>=12.0
python-socketio>=5.0.0
redis>=5.0.0
celery>=5.3.0
python-dotenv>=1.0.0
